move test scripts to enclosing directory
