+++ src/assets/testPresets.js
@@ -7,9 +7,9 @@ const fs = require('fs/promises');
 const path = require('path');
 // eslint-disable-next-line import/no-extraneous-dependencies
 const yaml = require('js-yaml');
-const { requireTS } = require('../../utils/require-ts.js');
+const { requireTS } = require('../utils/require-ts.js');
 
-const { templateTransform } = requireTS(path.join(__dirname, './templateTransform.js'));
+const { templateTransform } = requireTS(path.join(__dirname, './presetdata/templateTransform.js'));
 
 const directory = './src/assets/presetdata/';
